CT-177 Yeti filepath backslashes
Yeti nodes may store filepaths in either forward-slash or back-slash
format. This is problematic if the content is initially created on one
platform, like Windows, and saves the paths with backslashes.  But is
then subsequently opened on a different platform, like linux, which
doesn't acknowledge backslashes as path separators (since a backslash is
a valid character in unix path).  Though yeti handles this just fine, it
confuses our client tools.  
This fix chooses between two evils and simply assumes that any
backslashes are intended to be directory separators and converts them to
forward slashes (which I imagine is what the yeti plugin must be doing
too).